+++ src/main/java/com/monkey/web/controller/dtos/DeviceProductInput.java
@@ -1,9 +1,8 @@
 package com.monkey.web.controller.dtos;
 
-import java.math.BigDecimal;
 
 public class DeviceProductInput {
-    public BigDecimal price;
+    public Integer price;
     public String id;
     public Integer productId;
     public  Integer isSale;
