+++ src/main/java/com/monkey/web/controller/dtos/OrderInput.java
@@ -4,6 +4,6 @@ import java.math.BigDecimal;
 
 public class OrderInput {
     public Integer productId;
-    public BigDecimal price;
+    public Integer price;
     public String deviceNum;
 }
